Changing things around..
I have moved the social links to the top to make it more appealing for newcomers.
I have also added a new background that is not the butterfly image. This is a more darker background to fit the dark theme.
Additionally, I reverted the yellow text back to grey, and made some updates in css.